Update Address interface to general
