+++ src/interfaces/general.ts
@@ -1,0 +1,9 @@
+interface Address {
+    street: string,
+    houseNumber: string,
+    postalCode: string,
+    city: string,
+    country: string
+}
+
+export { Address };
