+++ src/interfaces/people.ts
@@ -1,0 +1,13 @@
+import { Address } from "./general";
+
+interface People {
+    name: string,
+    dob: Date,
+    address: Address,
+    vatId: string,
+    ssn: string,
+    email: string,
+    phoneNumber: string
+}
+
+export { People };
